Ignore sigterm coming from anything except auditd
Only honor the signal if it comes from the parent process so that other
tasks (cough, systemctl, cough) can't make the plugin exit without
the dispatcher in agreement. Otherwise it will restart the plugin.

+++ audisp/plugins/filter/audisp-filter.c
@@ -397,9 +397,15 @@ static void child_handler(int sig)
 
 /*
  * SIGTERM handler
+ *
+ * Only honor the signal if it comes from the parent process so that other
+ * tasks (cough, systemctl, cough) can't make the plugin exit without
+ * the dispatcher in agreement. Otherwise it will restart the plugin.
  */
-static void term_handler(int sig)
+static void term_handler(int sig, siginfo_t *info, void *ucontext)
 {
+	if (info && info->si_pid != getppid())
+		return;
 	kill(cpid, sig);
 	stop = 1;
 	auplugin_stop();
@@ -458,12 +464,13 @@ int main(int argc, const char* argv[])
 	sa.sa_flags = 0;
 	sigemptyset(&sa.sa_mask);
 	/* Set handler for the ones we care about */
-	sa.sa_handler = term_handler;
-	sigaction(SIGTERM, &sa, NULL);
 	sa.sa_handler = hup_handler;
 	sigaction(SIGHUP, &sa, NULL);
 	sa.sa_handler = child_handler;
 	sigaction(SIGCHLD, &sa, NULL);
+	sa.sa_sigaction = term_handler;
+	sa.sa_flags = SA_SIGINFO;
+	sigaction(SIGTERM, &sa, NULL);
 
 #ifdef HAVE_LIBCAP_NG
 	// Drop capabilities
